backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
rasterio==1.3.10
numpy==1.26.4
torch==2.3.0
torchvision==0.18.0
transformers==4.41.0
celery==5.4.0
redis==5.0.4
websockets==12.0
shapely==2.0.4
pyproj==3.6.1
geojson==3.1.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
python-dotenv==1.0.1
httpx==0.27.0
